requirements.txt
argcomplete==1.12.3
asgiref==3.5.2
attrs==21.4.0
aws==0.2.5
bcrypt==3.2.0
beautifulsoup4==4.11.1
boto==2.49.0
boto3==1.20.46
botocore==1.23.46
cachetools==5.0.0
capstone==5.0.0rc2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
ci-info==0.3.0
click==6.6
colorama==0.4.4
colored-traceback==0.3.0
configobj==5.0.6
configparser==5.2.0
crcmod==1.7
cryptography==3.4.7
Deprecated==1.2.13
distro==1.0.1
Django==4.0.5
dnspython==2.1.0
etelemetry==0.3.0
fabric==2.6.0
fasteners==0.16.2
filelock==3.7.1
fitz==0.0.1.dev2
Flask==1.0.2
Flask-MySQLdb==1.0.1
flask-unsign==1.2.0
future==0.18.2
gcloud==0.17.0
gcs-oauth2-boto-plugin==2.7
gevent==21.12.0
google-api-core==2.5.0
google-apitools==0.5.32
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-reauth==0.1.1
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
greenlet==1.1.2
grequests==0.6.0
gsutil==4.62
httplib2==0.19.1
idna==2.8
iniconfig==1.1.1
intervaltree==3.1.0
invoke==1.5.0
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==2.8
jmespath==0.10.0
jwcrypto==1.0
jws==0.1.3
looseversion==1.0.1
lxml==4.9.0
Mako==1.2.0
MarkupSafe==2.0.1
mergedeep==1.3.4
mock==2.0.0
monotonic==1.6
mysql-connector==2.2.9
mysqlclient==2.1.0
networkx==2.8.5
nibabel==4.0.1
nipype==1.8.3
numpy==1.23.1
oauth2client==3.0.0
packaging==21.3
pandas==1.4.3
paramiko==2.7.2
pathlib==1.0.1
pathlib2==2.3.5
pbr==5.6.0
pluggy==1.0.0
plumbum==1.7.2
prettytable==2.1.0
protobuf==4.21.4
prov==2.0.0
psutil==5.9.0
pwn==1.0
pwntools==4.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.20
pycryptodome==3.4.3
pycurl==7.44.1
pydot==1.4.2
pyelftools==0.28
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.2
PyJWT==2.4.0
pymongo==3.11.4
PyMuPDF==1.20.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyrebase==3.0.27
pyserial==3.5
pyserv==1.1.1
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.1
pyu2f==0.1.5
pyxnat==1.4
PyYAML==5.4.1
rdflib==6.2.0
requests==2.21.0
requests-toolbelt==0.7.0
retry-decorator==1.1.1
ROPGadget==6.7
rpyc==5.1.0
rsa==4.7.2
s3recon==1.4.0
S3Scanner==2.0.2
s3transfer==0.5.0
scipy==1.9.0
secret==0.8
Serv==0.3.0
sh==1.11
simplejson==3.8.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tabulate==0.8.9
tk==0.1.0
tld==0.12.6
toml==0.10.2
traits==6.3.2
unicorn==1.0.3
uploadserver==2.0.1
urllib3==1.24.3
wcwidth==0.2.5
Werkzeug==2.0.2
wfuzz==3.1.0
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0
